refactor: don't import subp function directly (#5065)
When subp function is imported directly, it cannot be easily mocked in
unit tests. Update imports and tests accordingly.
