+++ cloudinit/config/cc_ansible.py
@@ -9,12 +9,12 @@ from copy import deepcopy
 from textwrap import dedent
 from typing import Optional
 
+from cloudinit import subp
 from cloudinit.cloud import Cloud
 from cloudinit.config import Config
 from cloudinit.config.schema import MetaSchema, get_meta_doc
 from cloudinit.distros import ALL_DISTROS, Distro
 from cloudinit.settings import PER_INSTANCE
-from cloudinit.subp import subp, which
 from cloudinit.util import Version, get_cfg_by_path
 
 meta: MetaSchema = {
@@ -100,7 +100,7 @@ class AnsiblePull(abc.ABC):
         return self.distro.do_as(command, self.run_user, **kwargs)
 
     def subp(self, command, **kwargs):
-        return subp(command, update_env=self.env, **kwargs)
+        return subp.subp(command, update_env=self.env, **kwargs)
 
     @abc.abstractmethod
     def is_installed(self):
@@ -165,7 +165,7 @@ class AnsiblePullDistro(AnsiblePull):
             self.distro.install_packages([pkg_name])
 
     def is_installed(self) -> bool:
-        return bool(which("ansible"))
+        return bool(subp.which("ansible"))
 
 
 def handle(name: str, cfg: Config, cloud: Cloud, args: list) -> None:
